chore: ThemeCard 컴포넌트를 공용 컴포넌트 파일로 이동
